refactor: improve folder structure consistency
- Move Post enums (PostSorting, Visibility) to entity/enums folder
- Remove unused MissionHandler.java
- Move UserWithdrawalScheduler to user/scheduler folder
- Flatten Battle DTO folder structure (remove nested BattleCmtDTO, BattleDTO, BattleLikeDTO folders)
- Flatten Battle Service folder structure (remove nested BattleService, BattleCmtService, BattleLikeService folders)
- Add converter folders to emailtoken and report domains
- Consolidate S3 files to global/infra/s3 folder
