+++ src/main/java/UMC_7th/Closit/domain/battle/controller/BattleController.java
@@ -1,13 +1,13 @@
 package UMC_7th.Closit.domain.battle.controller;
 
 import UMC_7th.Closit.domain.battle.converter.BattleConverter;
-import UMC_7th.Closit.domain.battle.dto.BattleDTO.BattleRequestDTO;
-import UMC_7th.Closit.domain.battle.dto.BattleDTO.BattleResponseDTO;
+import UMC_7th.Closit.domain.battle.dto.BattleRequestDTO;
+import UMC_7th.Closit.domain.battle.dto.BattleResponseDTO;
 import UMC_7th.Closit.domain.battle.entity.*;
 import UMC_7th.Closit.domain.battle.entity.enums.BattleSorting;
 import UMC_7th.Closit.domain.battle.entity.enums.BattleStatus;
-import UMC_7th.Closit.domain.battle.service.BattleService.BattleCommandService;
-import UMC_7th.Closit.domain.battle.service.BattleService.BattleQueryService;
+import UMC_7th.Closit.domain.battle.service.BattleCommandService;
+import UMC_7th.Closit.domain.battle.service.BattleQueryService;
 import UMC_7th.Closit.domain.user.entity.User;
 import UMC_7th.Closit.global.apiPayload.ApiResponse;
 import UMC_7th.Closit.security.SecurityUtil;
