+++ src/main/java/UMC_7th/Closit/domain/battle/converter/BattleConverter.java
@@ -1,7 +1,7 @@
 package UMC_7th.Closit.domain.battle.converter;
 
-import UMC_7th.Closit.domain.battle.dto.BattleDTO.BattleRequestDTO;
-import UMC_7th.Closit.domain.battle.dto.BattleDTO.BattleResponseDTO;
+import UMC_7th.Closit.domain.battle.dto.BattleRequestDTO;
+import UMC_7th.Closit.domain.battle.dto.BattleResponseDTO;
 import UMC_7th.Closit.domain.battle.entity.*;
 import UMC_7th.Closit.domain.battle.entity.enums.BattleStatus;
 import UMC_7th.Closit.domain.battle.entity.enums.ChallengeStatus;
