+++ src/main/java/UMC_7th/Closit/domain/battle/dto/BattleCommentResponseDTO.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.dto.BattleLikeDTO;
+package UMC_7th.Closit.domain.battle.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
@@ -9,15 +9,17 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class BattleLikeResponseDTO {
+public class BattleCommentResponseDTO {
 
     @Builder
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class CreateBattleLikeResultDTO { // 배틀 좋아요 생성
-        private Long battleLikeId;
+    public static class CreateBattleCommentDTO { // 배틀 댓글 생성
+        private Long battleCommentId;
+        private Long parentBattleCommentId;
         private String clositId;
+        private String thumbnail;
         @JsonFormat(pattern = "yyyy/MM/dd HH:mm:ss")
         private LocalDateTime createdAt;
     }
@@ -26,9 +28,13 @@ public class BattleLikeResponseDTO {
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class BattleLikePreviewDTO { // 배틀 좋아요 조회
-        private Long battleLikeId;
+    public static class BattleCommentPreviewDTO { // 배틀 댓글 조회
+        private Long battleCommentId;
+        private Long parentBattleCommentId;
         private String clositId;
+        private String thumbnail;
+        private String content;
+        private List<BattleCommentPreviewDTO> childrenBattleComments;
         @JsonFormat(pattern = "yyyy/MM/dd HH:mm:ss")
         private LocalDateTime createdAt;
     }
@@ -37,8 +43,8 @@ public class BattleLikeResponseDTO {
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class BattleLikePreviewListDTO {
-        private List<BattleLikePreviewDTO> battleLikePreviewDTOList;
+    public static class BattleCommentPreviewListDTO {
+        private List<BattleCommentPreviewDTO> battleCommentPreviewList;
         private Integer listSize;
         private boolean isFirst;
         private boolean isLast;
