+++ src/main/java/UMC_7th/Closit/domain/battle/dto/BattleLikeResponseDTO.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.dto.BattleCmtDTO;
+package UMC_7th.Closit.domain.battle.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
@@ -9,17 +9,15 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class BattleCommentResponseDTO {
+public class BattleLikeResponseDTO {
 
     @Builder
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class CreateBattleCommentDTO { // 배틀 댓글 생성
-        private Long battleCommentId;
-        private Long parentBattleCommentId;
+    public static class CreateBattleLikeResultDTO { // 배틀 좋아요 생성
+        private Long battleLikeId;
         private String clositId;
-        private String thumbnail;
         @JsonFormat(pattern = "yyyy/MM/dd HH:mm:ss")
         private LocalDateTime createdAt;
     }
@@ -28,13 +26,9 @@ public class BattleCommentResponseDTO {
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class BattleCommentPreviewDTO { // 배틀 댓글 조회
-        private Long battleCommentId;
-        private Long parentBattleCommentId;
+    public static class BattleLikePreviewDTO { // 배틀 좋아요 조회
+        private Long battleLikeId;
         private String clositId;
-        private String thumbnail;
-        private String content;
-        private List<BattleCommentPreviewDTO> childrenBattleComments;
         @JsonFormat(pattern = "yyyy/MM/dd HH:mm:ss")
         private LocalDateTime createdAt;
     }
@@ -43,8 +37,8 @@ public class BattleCommentResponseDTO {
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class BattleCommentPreviewListDTO {
-        private List<BattleCommentPreviewDTO> battleCommentPreviewList;
+    public static class BattleLikePreviewListDTO {
+        private List<BattleLikePreviewDTO> battleLikePreviewDTOList;
         private Integer listSize;
         private boolean isFirst;
         private boolean isLast;
