+++ src/main/java/UMC_7th/Closit/domain/battle/dto/BattleRequestDTO.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.dto.BattleDTO;
+package UMC_7th.Closit.domain.battle.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
