+++ src/main/java/UMC_7th/Closit/domain/battle/service/BattleCmtCommandService.java
@@ -1,6 +1,6 @@
-package UMC_7th.Closit.domain.battle.service.BattleCmtService;
+package UMC_7th.Closit.domain.battle.service;
 
-import UMC_7th.Closit.domain.battle.dto.BattleCmtDTO.BattleCommentRequestDTO;
+import UMC_7th.Closit.domain.battle.dto.BattleCommentRequestDTO;
 import UMC_7th.Closit.domain.battle.entity.BattleComment;
 
 public interface BattleCmtCommandService {
