+++ src/main/java/UMC_7th/Closit/domain/battle/service/BattleCmtQueryService.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.service.BattleCmtService;
+package UMC_7th.Closit.domain.battle.service;
 
 import UMC_7th.Closit.domain.battle.entity.BattleComment;
 import org.springframework.data.domain.Slice;
