+++ src/main/java/UMC_7th/Closit/domain/battle/service/BattleLikeQueryService.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.service.BattleLikeService;
+package UMC_7th.Closit.domain.battle.service;
 
 import UMC_7th.Closit.domain.battle.entity.BattleLike;
 import org.springframework.data.domain.Slice;
