+++ src/main/java/UMC_7th/Closit/domain/post/controller/PostController.java
@@ -4,13 +4,13 @@ import UMC_7th.Closit.domain.post.converter.PostConverter;
 import UMC_7th.Closit.domain.post.dto.PostRequestDTO;
 import UMC_7th.Closit.domain.post.dto.PostResponseDTO;
 import UMC_7th.Closit.domain.post.entity.Post;
-import UMC_7th.Closit.domain.post.entity.Visibility;
-import UMC_7th.Closit.domain.post.entity.PostSorting;
+import UMC_7th.Closit.domain.post.entity.enums.Visibility;
+import UMC_7th.Closit.domain.post.entity.enums.PostSorting;
 import UMC_7th.Closit.domain.post.service.PostCommandService;
 import UMC_7th.Closit.domain.post.service.PostQueryService;
 import UMC_7th.Closit.domain.user.entity.User;
 import UMC_7th.Closit.global.apiPayload.ApiResponse;
-import UMC_7th.Closit.global.s3.S3Service;
+import UMC_7th.Closit.global.infra.s3.S3Service;
 import UMC_7th.Closit.security.SecurityUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
