+++ src/main/java/UMC_7th/Closit/domain/post/dto/PostRequestDTO.java
@@ -1,6 +1,6 @@
 package UMC_7th.Closit.domain.post.dto;
 
-import UMC_7th.Closit.domain.post.entity.Visibility;
+import UMC_7th.Closit.domain.post.entity.enums.Visibility;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
