+++ src/main/java/UMC_7th/Closit/domain/post/dto/PostResponseDTO.java
@@ -1,6 +1,6 @@
 package UMC_7th.Closit.domain.post.dto;
 
-import UMC_7th.Closit.domain.post.entity.Visibility;
+import UMC_7th.Closit.domain.post.entity.enums.Visibility;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
