+++ src/main/java/UMC_7th/Closit/domain/post/entity/enums/Visibility.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.post.entity;
+package UMC_7th.Closit.domain.post.entity.enums;
 
 public enum Visibility {
     PUBLIC, FRIEND, PRIVATE;
