+++ src/main/java/UMC_7th/Closit/domain/post/service/PostQueryServiceImpl.java
@@ -6,10 +6,7 @@ import UMC_7th.Closit.domain.highlight.repository.HighlightRepository;
 import UMC_7th.Closit.domain.post.converter.PostConverter;
 import UMC_7th.Closit.domain.post.dto.PostResponseDTO;
 import UMC_7th.Closit.domain.post.entity.*;
-import UMC_7th.Closit.domain.post.entity.Hashtag;
-import UMC_7th.Closit.domain.post.entity.ItemTag;
-import UMC_7th.Closit.domain.post.entity.Post;
-import UMC_7th.Closit.domain.post.entity.Visibility;
+import UMC_7th.Closit.domain.post.entity.enums.Visibility;
 import UMC_7th.Closit.domain.post.repository.*;
 import UMC_7th.Closit.domain.user.entity.User;
 import UMC_7th.Closit.global.apiPayload.code.status.ErrorStatus;
