--- src/main/java/UMC_7th/Closit/global/apiPayload/exception/handler/MissionHandler.java
@@ -1,9 +1,0 @@
-package UMC_7th.Closit.global.apiPayload.exception.handler;
-
-import UMC_7th.Closit.global.apiPayload.code.BaseErrorCode;
-import UMC_7th.Closit.global.apiPayload.exception.GeneralException;
-
-public class MissionHandler extends GeneralException {
-
-    public MissionHandler(BaseErrorCode errorCode) { super(errorCode); }
-}
