+++ src/main/java/UMC_7th/Closit/global/infra/s3/MultipartJackson2HttpMessageConverter.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.global.s3;
+package UMC_7th.Closit.global.infra.s3;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.MediaType;
